Move site info card related classes to cards package
